+++ common/list/presentation/src/commonMain/kotlin/ru/andvl/mppapp/list/models/DelaAction.kt
@@ -1,5 +1,5 @@
 package ru.andvl.mppapp.list.models
 
 sealed interface DelaAction {
-    data class OpenDetails(val id: Int) : DelaAction
+    data class OpenDetails(val id: Long) : DelaAction
 }
