+++ common/list/presentation/src/commonMain/kotlin/ru/andvl/mppapp/list/models/DelaEvent.kt
@@ -3,5 +3,5 @@ package ru.andvl.mppapp.list.models
 sealed interface DelaEvent {
     object Refresh : DelaEvent
     object DeloOpen : DelaEvent
-    data class DeloClick(val id: Int) : DelaEvent
+    data class DeloClick(val id: Long) : DelaEvent
 }
